+++ src/renderer/components/Ledger/LedgerTab.vue
@@ -1,0 +1,105 @@
+<template >
+<el-row>
+    <p>Ledgers:</p>
+    <el-collapse v-model="expanded_ledger_items">
+    <div v-for="(ledger, key) in ledgers" :key="key">
+    <el-collapse-item v-bind:title="'Name: ' + ledger.name" :name="key">
+    <el-row>
+    <div>
+    <vue-json-pretty
+    :deep=1
+    :data="ledger">
+    </vue-json-pretty>
+    <el-form class="ledger-form">
+        <el-form-item >
+            <el-button type="primary" @click="removeLedger(ledger)">delete</el-button>
+        </el-form-item>
+        <el-button v-on:click="collapse_expanded_ledger_item(key)">^</el-button>
+        </el-form>
+            </div>
+            </el-row>
+            </el-collapse-item>
+            </div>
+            </el-collapse>
+            <p>Add new ledger:</p>
+            <el-form :model=ledger_form>
+            <el-form-group >
+            <span slot="label">Name:</span>
+            <el-input v-model="ledger_form.name" style="width:100px;"> </el-input>
+            <span slot="label">Url to genesis file:</span>
+            <el-input v-model="ledger_form.gen_url" style="width:100px;"> </el-input>
+            </el-form-group>
+            <el-form-item>
+            <el-button type="primary" @click="addLedger()">add new ledger</el-button>
+        </el-form-item>
+    </el-form>
+</el-row>
+</template>
+
+<script>
+import VueJsonPretty from 'vue-json-pretty';
+const { clipboard } = require('electron');
+import VueQrcode from '@chenfengyuan/vue-qrcode';
+
+export default {
+  name: 'ledger-tab',
+  message_bus: 'derive',
+  components: {
+    VueJsonPretty,
+  },
+  data () {
+    return {
+        ledgers:{
+            '12345234':{
+            'id':'12345234',
+            'name':'sov',
+            },
+            '22345234':{
+            'id':'22345234',
+            'name':'von',
+            },
+            '32345234':{
+            'id':'32345234',
+            'name':'bank',
+            },
+            '42345234':{
+            'id':'42345234',
+            'name':'adams',
+            },
+        },
+        ledger_form:{
+            'name':'',
+            'gen_url':'',
+        },
+        expanded_ledger_items:[],
+    }
+  },
+  created: function() {
+    let component = this; // Safe rerefence to this
+    this.$message_bus.$on(
+      'did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/admin-connections/1.0/invitation',
+      msg => component.fetchAgentInvitations()
+    );
+    this.$message_bus.$on('invitations', () => component.fetchAgentInvitations());
+  },
+  methods: {
+    async collapse_expanded_ledger_item(id){
+      let index = this.expanded_ledger_items.indexOf(id)
+      this.expanded_ledger_items.splice(index, 1);
+    },
+  },
+  computed: {
+    ledgerSchemas(ledger_name){
+    console.log("ledger name:",ledger_name)
+    let schemas = Object.keys(this.schemas).reduce((acc,key)=>
+    ("ledgers" in this.schemas[key] && ledger_name in this.schemas[key].ledgers ?{
+        ...acc,
+        [key]: this.schemas[key]
+    } : acc
+    ), {})
+    console.log("schemas",schemas)
+    return schemas
+},
+  }
+}
+</script>
